refactor: extract PendingWithdrawals component
